+++ core/tools/renderQueue/render_submit.py
@@ -6,10 +6,11 @@
 # (c) 2015-2017 Gramercy Park Studios
 #
 # Batch Render Submitter
-# A UI for creating render jobs to send to the render queue.
+# A UI for creating render jobs to send to a render queue manager.
 
 
 import os
+import subprocess
 import sys
 import time
 
@@ -36,20 +37,18 @@ WINDOW_OBJECT = "renderSubmitUI"
 UI_FILE = "render_submit_ui.ui"
 STYLESHEET = "style.qss"  # Set to None to use the parent app's stylesheet
 
-# Other options
-DOCK_WITH_MAYA_UI = False
-DOCK_WITH_NUKE_UI = False
+CREATE_NO_WINDOW = 0x08000000
 
 
 # ----------------------------------------------------------------------------
 # Main dialog class
 # ----------------------------------------------------------------------------
 
-class renderSubmitDialog(QtWidgets.QDialog):
+class RenderSubmitUI(QtWidgets.QMainWindow):
 	""" Main dialog class.
 	"""
 	def __init__(self, parent=None):
-		super(renderSubmitDialog, self).__init__(parent)
+		super(RenderSubmitUI, self).__init__(parent)
 
 		# Set object name and window title
 		self.setObjectName(WINDOW_OBJECT)
@@ -62,14 +61,28 @@ class renderSubmitDialog(QtWidgets.QDialog):
 			with open(qss, "r") as fh:
 				self.ui.setStyleSheet(fh.read())
 
+		# Set the main widget
+		self.setCentralWidget(self.ui)
+
 		# Set window flags
-		self.setWindowFlags(QtCore.Qt.Dialog)
-		self.ui.setWindowFlags(QtCore.Qt.CustomizeWindowHint | 
-			                   QtCore.Qt.WindowTitleHint)
+		self.setWindowFlags(QtCore.Qt.Tool)
+		# self.setWindowFlags(QtCore.Qt.Dialog)
+		# self.ui.setWindowFlags(QtCore.Qt.CustomizeWindowHint | 
+		# 	                   QtCore.Qt.WindowTitleHint)
 
 		# Connect signals & slots
+		self.ui.submitTo_comboBox.currentIndexChanged.connect(self.setQueueManagerFromComboBox)
 		self.ui.type_comboBox.currentIndexChanged.connect(self.setJobTypeFromComboBox)
 		self.ui.sceneBrowse_toolButton.clicked.connect(self.sceneBrowse)
+
+		self.ui.pool_comboBox.currentIndexChanged.connect(self.storeComboBoxValue)
+		self.ui.group_comboBox.currentIndexChanged.connect(self.storeComboBoxValue)
+		self.ui.pool_comboBox.editTextChanged.connect(self.storeComboBoxValue)
+		self.ui.group_comboBox.editTextChanged.connect(self.storeComboBoxValue)
+		self.ui.getPools_toolButton.clicked.connect(self.getDeadlinePools)
+		self.ui.getGroups_toolButton.clicked.connect(self.getDeadlineGroups)
+		self.ui.priority_spinBox.valueChanged.connect(self.storeSpinBoxValue)
+		self.ui.comment_lineEdit.textEdited.connect(self.storeLineEditValue)
 
 		self.ui.frameRange_lineEdit.editingFinished.connect(self.calcFrameList)  # was textEdited
 		self.ui.taskSize_spinBox.valueChanged.connect(self.calcFrameList)
@@ -77,8 +90,25 @@ class renderSubmitDialog(QtWidgets.QDialog):
 		self.ui.submit_pushButton.clicked.connect(self.submit)
 		self.ui.close_pushButton.clicked.connect(self.exit)
 
+		# Context menus
+		self.ui.frameListOptions_toolButton.setContextMenuPolicy(QtCore.Qt.ActionsContextMenu)
+
+		self.actionFrameListOption1 = QtWidgets.QAction("Shot default", None)
+		self.actionFrameListOption1.triggered.connect(self.setFrameListPreset)
+		self.ui.frameListOptions_toolButton.addAction(self.actionFrameListOption1)
+
+		self.actionFrameListOption2 = QtWidgets.QAction("Timeline", None)
+		self.actionFrameListOption2.triggered.connect(self.setFrameListPreset)
+		self.ui.frameListOptions_toolButton.addAction(self.actionFrameListOption2)
+
+		self.actionFrameListOption3 = QtWidgets.QAction("First, last, sequential", None)
+		self.actionFrameListOption3.triggered.connect(self.setFrameListPreset)
+		self.ui.frameListOptions_toolButton.addAction(self.actionFrameListOption3)
+
+		self.ui.frameListOptions_toolButton.setEnabled(False)  # TEMP DISABLE
+
 		# Set input validators
-		frame_list_validator = QtGui.QRegExpValidator( QtCore.QRegExp(r'[\d\-, ]+'), self.ui.frameRange_lineEdit)
+		frame_list_validator = QtGui.QRegExpValidator(QtCore.QRegExp(r'[\d\-, ]+'), self.ui.frameRange_lineEdit)
 		self.ui.frameRange_lineEdit.setValidator(frame_list_validator)
 
 		# Instantiate render queue class and load data
@@ -94,13 +124,25 @@ class renderSubmitDialog(QtWidgets.QDialog):
 		# Read user prefs config file - if it doesn't exist it will be created
 		userPrefs.read()
 
+		# Restore widgets from user prefs settings
+		self.submitTo = userPrefs.query('rendersubmit', 'submitto', default=self.ui.submitTo_comboBox.currentText())
+		self.ui.submitTo_comboBox.setCurrentIndex(self.ui.submitTo_comboBox.findText(self.submitTo))
+
+		pool = userPrefs.query('rendersubmit', 'pool', datatype='str', default='')
+		self.ui.pool_comboBox.addItem(pool)
+
+		group = userPrefs.query('rendersubmit', 'group', datatype='str', default='')
+		self.ui.group_comboBox.addItem(group)
+
+		priority = userPrefs.query('rendersubmit', 'priority', datatype='int', default='')
+		self.ui.priority_spinBox.setValue(priority)
+
+		comment = userPrefs.query('rendersubmit', 'comment', datatype='str', default='')
+		self.ui.comment_lineEdit.setText(comment)
+
 		# Set job type from Icarus environment when possible
 		if os.environ['IC_ENV'] == 'STANDALONE':
-			try:
-				self.jobType = userPrefs.config.get('renderqueue', 'lastrenderjobtype')
-			except:
-				self.jobType = self.ui.type_comboBox.currentText()
-
+			self.jobType = userPrefs.query('rendersubmit', 'lastrenderjobtype', default=self.ui.type_comboBox.currentText())
 			self.ui.type_comboBox.setCurrentIndex(self.ui.type_comboBox.findText(self.jobType))
 			self.setJobType()
 			self.setSceneList()
@@ -123,6 +165,8 @@ class renderSubmitDialog(QtWidgets.QDialog):
 				self.ui.scene_comboBox.addItem(msg)
 				self.ui.submit_pushButton.setToolTip(msg)
 				self.ui.submit_pushButton.setEnabled(False)
+
+			self.getRenderers()
 
 			self.ui.render_groupBox.setEnabled(False)
 			self.ui.sceneBrowse_toolButton.hide()
@@ -160,9 +204,20 @@ class renderSubmitDialog(QtWidgets.QDialog):
 			self.ui.flags_groupBox.setChecked(True)
 			self.ui.flags_lineEdit.setText(flags)
 
-		self.ui.show()
-		self.ui.raise_()
+		# self.ui.show()
+		# self.ui.raise_()
+		self.show()
+		self.raise_()
 		return self.returnValue
+
+
+	def setQueueManagerFromComboBox(self):
+		""" Set queue manager to submit job to - called when the job type
+			combo box value is changed.
+		"""
+		self.submitTo = self.ui.submitTo_comboBox.currentText()
+		userPrefs.edit('rendersubmit', 'submitto', self.submitTo)
+		self.ui.submit_pushButton.setText("Submit to %s" %self.submitTo)
 
 
 	def setJobTypeFromComboBox(self):
@@ -170,7 +225,7 @@ class renderSubmitDialog(QtWidgets.QDialog):
 			changed.
 		"""
 		self.jobType = self.ui.type_comboBox.currentText()
-		userPrefs.edit('renderqueue', 'lastrenderjobtype', self.jobType)
+		userPrefs.edit('rendersubmit', 'lastrenderjobtype', self.jobType)
 		self.setJobType()
 		self.setSceneList()
 
@@ -185,12 +240,16 @@ class renderSubmitDialog(QtWidgets.QDialog):
 			except KeyError:
 				self.relativeScenesDir = ""
 			self.ui.scene_label.setText("Scene:")
+			self.ui.renderer_label.show()
+			self.ui.renderer_comboBox.show()
 		elif self.jobType == 'Nuke':
 			try:
 				self.relativeScenesDir = osOps.absolutePath( '%s/%s' %(os.environ['NUKEDIR'], 'scripts') )
 			except KeyError:
 				self.relativeScenesDir = ""
 			self.ui.scene_label.setText("Script:")
+			self.ui.renderer_label.hide()
+			self.ui.renderer_comboBox.hide()
 
 		self.relativeScenesToken = '...'  # Representative string to replace the path specified above
 
@@ -257,6 +316,47 @@ class renderSubmitDialog(QtWidgets.QDialog):
 				verbose.print_("Warning: Only %s belonging to the current shot can be submitted." %fileTerminology, 2)
 
 
+	# def storeOptions(self):
+	# 	""" Store persistent settings in user prefs.
+	# 	"""
+	# 	self.storeValue('pool', self.ui.pool_comboBox.currentText())
+	# 	self.storeValue('group', self.ui.group_comboBox.currentText())
+
+
+	# @QtCore.Slot()
+	def storeSpinBoxValue(self):
+		""" Get the value from a Spin Box and store in XML data.
+		"""
+		attr = self.sender().property('xmlTag')
+		value = self.sender().value()
+		self.storeValue(attr, value)
+
+
+	# @QtCore.Slot()
+	def storeLineEditValue(self):
+		""" Get the value from a Line Edit and store in XML data.
+		"""
+		attr = self.sender().property('xmlTag')
+		value = self.sender().text()
+		self.storeValue(attr, value)
+
+
+	# @QtCore.Slot()
+	def storeComboBoxValue(self):
+		""" Get the value from a Combo Box and store in XML data.
+		"""
+		attr = self.sender().property('xmlTag')
+		value = self.sender().currentText()
+		self.storeValue(attr, value)
+
+
+	def storeValue(self, attr, value=""):
+		""" Store value in XML data.
+		"""
+		verbose.print_("%20s %s.%s=%s" %(type(value), 'rendersubmit', attr, value), 4)
+		userPrefs.edit('rendersubmit', attr, value)
+
+
 	def getFrameList(self):
 		""" Get frame range from shot settings.
 		"""
@@ -267,6 +367,66 @@ class renderSubmitDialog(QtWidgets.QDialog):
 			self.ui.frameRange_lineEdit.setText("")
 			self.ui.overrideFrameRange_groupBox.setChecked(False)
 
+
+	def getDeadlinePools(self):
+		""" Get Deadline pools & populate combo box.
+		"""
+		pools = subprocess.check_output([os.environ['DEADLINECMDVERSION'], '-pools'], creationflags=CREATE_NO_WINDOW)
+		self.populateComboBox(self.ui.pool_comboBox, self.strToList(pools))
+
+
+	def getDeadlineGroups(self):
+		""" Get Deadline groups & populate combo box.
+		"""
+		groups = subprocess.check_output([os.environ['DEADLINECMDVERSION'], '-groups'], creationflags=CREATE_NO_WINDOW)
+		self.populateComboBox(self.ui.group_comboBox, self.strToList(groups))
+
+
+	def getRenderers(self):
+		""" Get Maya renderers & populate combo box.
+		"""
+		renderers = self.getRenderer()
+		rendererLs = [renderers]
+		# verbose.print_(rendererLs)
+		self.populateComboBox(self.ui.renderer_comboBox, rendererLs)
+
+
+	def strToList(self, string):
+		""" Convert string to list by splitting into lines.
+			Encode bytes to string for Python 3.
+		"""
+		string = string.decode()
+		ls = string.splitlines()
+		verbose.print_(ls)
+		return ls
+
+
+	def populateComboBox(self, comboBox, contentsLs):
+		""" Populate combo box.
+		"""
+		# Store current value
+		current = comboBox.currentText()
+
+		# Clear menu
+		comboBox.clear()
+
+		# Populate menu
+		for item in contentsLs:
+			if item:
+				comboBox.addItem(item)
+
+		# Set to current value
+		index = comboBox.findText(current)
+		if index == -1:
+			comboBox.setCurrentIndex(0)
+		else:
+			comboBox.setCurrentIndex(index)
+
+
+	def setFrameListPreset(self):
+		""" Set frame list from preset
+		"""
+		pass
 
 
 	def calcFrameList(self, quiet=True):
@@ -302,18 +462,39 @@ class renderSubmitDialog(QtWidgets.QDialog):
 		return True
 					
 
-	def guessMayaProject(self, scene):
-		""" Try to guess the Maya project directory based on the scene name.
+	def getMayaProject(self, scene):
+		""" Get the Maya Project. If the environmnet variable is not set, try
+			to guess the Maya project directory based on the scene name.
 		"""
-		try:
-			mayaProject = os.environ['MAYADIR'].replace("\\", "/")  # Project is implicit if job is set
+		try:  # Project is implicit if job is set
+			mayaProject = os.environ['MAYADIR'].replace("\\", "/")
 		except KeyError:
 			mayaProject = scene.split('/scenes')[0]
 
 		return mayaProject
 
 
+	def getRenderer(self):
+		""" Get the current Maya renderer
+		"""
+		try:
+			renderer = mc.getAttr("defaultRenderGlobals.currentRenderer")
+		except KeyError:
+			renderer = ""
+
+		return renderer
+
+
 	def submit(self):
+		""" Submit job.
+		"""
+		if self.submitTo == "RenderQueue":
+			self.submitToRenderQueue()
+		if self.submitTo == "Deadline":
+			self.submitToDeadline()
+
+
+	def submitToRenderQueue(self):
 		""" Submit job to render queue.
 		"""
 		timeFormatStr = "%Y/%m/%d %H:%M:%S" # "%a, %d %b %Y %H:%M:%S"
@@ -351,12 +532,12 @@ class renderSubmitDialog(QtWidgets.QDialog):
 		if self.jobType == 'Maya':
 			renderCmdEnvVar = 'MAYARENDERVERSION'
 			mayaScene = self.absolutePath(self.ui.scene_comboBox.currentText()).replace("\\", "/")  # Implicit if submitting from Maya UI
-			mayaProject = self.guessMayaProject(mayaScene)
-			jobName = os.path.basename(mayaScene)
+			mayaProject = self.getMayaProject(mayaScene)
+			jobName = os.path.splitext(os.path.basename(mayaScene))[0]
 		elif self.jobType == 'Nuke':
 			renderCmdEnvVar = 'NUKEVERSION'
 			nukeScript = self.absolutePath(self.ui.scene_comboBox.currentText()).replace("\\", "/")  # Implicit if submitting from Nuke UI
-			jobName = os.path.basename(nukeScript)
+			jobName = os.path.splitext(os.path.basename(nukeScript))[0]
 
 		try:
 			renderCmd = os.environ[renderCmdEnvVar].replace("\\", "/")
@@ -397,6 +578,113 @@ class renderSubmitDialog(QtWidgets.QDialog):
 			return
 
 
+	def submitToDeadline(self):
+		""" Submit job to Deadline.
+		"""
+		# timeFormatStr = "%Y/%m/%d %H:%M:%S" # "%a, %d %b %Y %H:%M:%S"
+		pluginInfoFile = ""
+
+		if not self.calcFrameList(quiet=False):
+			return
+
+		###################
+		# Generic options #
+		###################
+
+		frames = self.ui.frameRange_lineEdit.text()
+		taskSize = self.ui.taskSize_spinBox.value()
+		framesMsg = '%d %s to be rendered; %d %s to be submitted.\n' %(len(self.numList), verbose.pluralise("frame", len(self.numList)), len(self.taskList), verbose.pluralise("task", len(self.taskList)))
+
+		pool = self.ui.pool_comboBox.currentText()
+		group = self.ui.group_comboBox.currentText()
+		priority = self.ui.priority_spinBox.value()
+		comment = self.ui.comment_lineEdit.text()
+
+		#############################
+		# Renderer-specific options #
+		#############################
+
+		if self.jobType == 'Maya':
+			plugin = 'MayaBatch'
+			version = os.environ['MAYA_VER']  #jobData.getAppVersion('Maya')
+			renderer = self.ui.renderer_comboBox.currentText()
+			scene = self.absolutePath(self.ui.scene_comboBox.currentText()).replace("\\", "/")
+			mayaProject = self.getMayaProject(scene)
+			jobName = os.path.splitext(os.path.basename(scene))[0]
+		elif self.jobType == 'Nuke':
+			plugin = 'Nuke'
+			scene = self.absolutePath(self.ui.scene_comboBox.currentText()).replace("\\", "/")
+			jobName = os.path.splitext(os.path.basename(scene))[0]
+
+		# Package option variables into tuples
+		genericOpts = jobName, self.jobType, frames, taskSize, priority
+		# if self.jobType == 'Maya':
+		# 	renderOpts = scene, mayaProject, flags, renderCmd
+		# elif self.jobType == 'Nuke':
+		# 	renderOpts = scene, flags, renderCmd
+
+		# Confirmation dialog
+		import pDialog
+
+		dialogTitle = 'Submit Render to Deadline - %s' %jobName
+		dialogMsg = ''
+		dialogMsg += 'Name:\t%s\nType:\t%s\nFrames:\t%s\nTask size:\t%s\nPriority:\t%s\n\n' %genericOpts
+		# if self.jobType == 'Maya':
+		# 	dialogMsg += 'Scene:\t%s\nProject:\t%s\nFlags:\t%s\nCommand:\t%s\n\n' %renderOpts
+		# elif self.jobType == 'Nuke':
+		# 	dialogMsg += 'Script:\t%s\nFlags:\t%s\nCommand:\t%s\n\n' %renderOpts
+		dialogMsg += framesMsg
+		dialogMsg += 'Do you want to continue?'
+
+		dialog = pDialog.dialog()
+		if dialog.display(dialogMsg, dialogTitle):
+
+			# Generate submission info files
+			jobInfoFile = os.path.splitext(scene)[0] + "_deadlineJobInfo.txt"
+			fh = open(jobInfoFile, 'w')
+			fh.write("Plugin=%s\n" %plugin)
+			fh.write("Name=%s\n" %jobName)
+			fh.write("Comment=%s\n" %comment)
+			fh.write("Frames=%s\n" %frames)
+			fh.write("ChunkSize=%s\n" %taskSize)
+			fh.write("Pool=%s\n" %pool)
+			fh.write("Group=%s\n" %group)
+			fh.write("Priority=%s\n" %priority)
+			if priority == 0:
+				fh.write("InitialStatus=Suspended\n")
+			# fh.write("OutputDirectory0=%s\n" %mayaOps.getRenderImagePath())
+			# fh.write("OutputFilename0=%s\n" %mayaOps.getRenderImageName())
+			fh.close()
+
+			pluginInfoFile = os.path.splitext(scene)[0] + "_deadlinePluginInfo.txt"
+			fh = open(pluginInfoFile, 'w')
+			fh.write("Version=%s\n" %version)
+			fh.write("Build=64bit\n")
+			fh.write("Renderer=%s\n" %renderer)
+			fh.write("StrictErrorChecking=1\n")
+			fh.write("ProjectPath=%s\n" %mayaProject)
+			# fh.write("OutputFilePath=%s\n" %mayaOps.getOutputFilePath())
+			# fh.write("OutputFilePrefix=%s\n" %mayaOps.getOutputFilePrefix())
+			fh.write("SceneFile=%s\n" %scene)
+			fh.close()
+
+			# Execute deadlinecommand
+			output = subprocess.check_output([os.environ['DEADLINECMDVERSION'], jobInfoFile, pluginInfoFile], creationflags=CREATE_NO_WINDOW)
+
+			# Delete submission info files
+			osOps.recurseRemove(jobInfoFile)
+			osOps.recurseRemove(pluginInfoFile)
+
+			# Post-confirmation dialog
+			dialogTitle = 'Submitted Render to Deadline - %s' %jobName
+			dialogMsg = ''
+			dialogMsg += 'Name:\t%s\nType:\t%s\nFrames:\t%s\nTask size:\t%s\nPriority:\t%s\n\n' %genericOpts
+			dialogMsg += output
+			dialog.display(dialogMsg, dialogTitle, conf=True)
+		else:
+			return
+
+
 	def showEvent(self, event):
 		""" Event handler for when window is shown.
 		"""
@@ -408,7 +696,8 @@ class renderSubmitDialog(QtWidgets.QDialog):
 	def exit(self):
 		""" Exit the dialog.
 		"""
-		self.ui.hide()
+		# self.ui.hide()
+		self.hide()
 
 # ----------------------------------------------------------------------------
 # End of main dialog class
@@ -418,12 +707,12 @@ class renderSubmitDialog(QtWidgets.QDialog):
 # def run_(**kwargs):
 # 	# for key, value in kwargs.iteritems():
 # 	# 	print "%s = %s" % (key, value)
-# 	renderSubmitDialog = renderSubmitDialog(**kwargs)
-# 	#renderSubmitDialog.setAttribute( QtCore.Qt.WA_DeleteOnClose )
-# 	print renderSubmitDialog
-# 	renderSubmitDialog.show()
-# 	#renderSubmitDialog.raise_()
-# 	#renderSubmitDialog.exec_()
+# 	renderSubmitUI = RenderSubmitUI(**kwargs)
+# 	#renderSubmitUI.setAttribute( QtCore.Qt.WA_DeleteOnClose )
+# 	print renderSubmitUI
+# 	renderSubmitUI.show()
+# 	#renderSubmitUI.raise_()
+# 	#renderSubmitUI.exec_()
 
 
 # ----------------------------------------------------------------------------
@@ -433,9 +722,9 @@ class renderSubmitDialog(QtWidgets.QDialog):
 def _maya_delete_ui():
 	""" Delete existing UI in Maya.
 	"""
-	if mc.window(WINDOW_OBJECT, q=True, exists=True):
+	if mc.window(WINDOW_OBJECT, query=True, exists=True):
 		mc.deleteUI(WINDOW_OBJECT)  # Delete window
-	if mc.dockControl('MayaWindow|' + WINDOW_TITLE, q=True, exists=True):
+	if mc.dockControl('MayaWindow|' + WINDOW_TITLE, query=True, exists=True):
 		mc.deleteUI('MayaWindow|' + WINDOW_TITLE)  # Delete docked window
 
 
@@ -450,7 +739,6 @@ def _nuke_delete_ui():
 def _maya_main_window():
 	""" Return Maya's main window.
 	"""
-#	for obj in QtWidgets.qApp.topLevelWidgets():
 	for obj in QtWidgets.QApplication.topLevelWidgets():
 		if obj.objectName() == 'MayaWindow':
 			return obj
@@ -460,98 +748,71 @@ def _maya_main_window():
 def _nuke_main_window():
 	""" Returns Nuke's main window.
 	"""
-#	for obj in QtWidgets.qApp.topLevelWidgets():
 	for obj in QtWidgets.QApplication.topLevelWidgets():
 		if (obj.inherits('QMainWindow') and obj.metaObject().className() == 'Foundry::UI::DockMainWindow'):
 			return obj
 	raise RuntimeError("Could not find DockMainWindow instance")
 
 
-def _nuke_set_zero_margins(widget_object):
-	""" Remove Nuke margins when docked UI.
-		More info:
-		https://gist.github.com/maty974/4739917
-	"""
-	parentApp = QtWidgets.QApplication.allWidgets()
-	parentWidgetList = []
-	for parent in parentApp:
-		for child in parent.children():
-			if widget_object.__class__.__name__ == child.__class__.__name__:
-				parentWidgetList.append(parent.parentWidget())
-				parentWidgetList.append(parent.parentWidget().parentWidget())
-				parentWidgetList.append(parent.parentWidget().parentWidget().parentWidget())
-
-				for sub in parentWidgetList:
-					for tinychild in sub.children():
-						try:
-							tinychild.setContentsMargins(0, 0, 0, 0)
-						except:
-							pass
-
-
 # ----------------------------------------------------------------------------
 # Run functions
 # ----------------------------------------------------------------------------
 
-def run_maya(frameRange=None, flags=None):
+def run_maya(**kwargs):
 	""" Run in Maya.
 	"""
+	# if mc.window(WINDOW_OBJECT, query=True, exists=True):
+	# 	mc.showWindow(WINDOW_OBJECT)
+	# else:
+	# 	_maya_delete_ui()  # Delete any already existing UI
+	# 	renderSubmitUI = RenderSubmitUI(parent=_maya_main_window())
+
+	# 	# Makes Maya perform magic which makes the window stay on top in OS X and
+	# 	# Linux. As an added bonus, it'll make Maya remember the window position.
+	# 	renderSubmitUI.setProperty("saveWindowPref", True)
+
+	# 	renderSubmitUI.display(**kwargs)  # Show the UI
+
 	_maya_delete_ui()  # Delete any already existing UI
-	renderSubmitUI = renderSubmitDialog(parent=_maya_main_window())
+	renderSubmitUI = RenderSubmitUI(parent=_maya_main_window())
 
 	# Makes Maya perform magic which makes the window stay on top in OS X and
 	# Linux. As an added bonus, it'll make Maya remember the window position.
 	renderSubmitUI.setProperty("saveWindowPref", True)
 
-	if not DOCK_WITH_MAYA_UI:
-		renderSubmitUI.display(frameRange=frameRange, flags=flags)  # Show the UI
-	elif DOCK_WITH_MAYA_UI:
-		allowed_areas = ['right', 'left']
-		mc.dockControl(WINDOW_TITLE, label=WINDOW_TITLE, area='left', content=WINDOW_OBJECT, allowedArea=allowed_areas)
+	renderSubmitUI.display(**kwargs)  # Show the UI
 
 
-def run_nuke():
+def run_nuke(**kwargs):
 	""" Run in Nuke.
 
 		Note:
 			If you want the UI to always stay on top, replace:
-			`renderSubmitUI.ui.setWindowFlags(QtCore.Qt.Tool)`
+			'renderSubmitUI.ui.setWindowFlags(QtCore.Qt.Tool)'
 			with:
-			`renderSubmitUI.ui.setWindowFlags(QtCore.Qt.WindowStaysOnTopHint)`
+			'renderSubmitUI.ui.setWindowFlags(QtCore.Qt.WindowStaysOnTopHint)'
 
 			If you want the UI to be modal:
-			`renderSubmitUI.ui.setWindowModality(QtCore.Qt.WindowModal)`
+			'renderSubmitUI.ui.setWindowModality(QtCore.Qt.WindowModal)'
 	"""
 	_nuke_delete_ui()  # Delete any already existing UI
-	if not DOCK_WITH_NUKE_UI:
-		renderSubmitUI = renderSubmitDialog(parent=_nuke_main_window())
-		renderSubmitUI.setWindowFlags(QtCore.Qt.Tool)
-		renderSubmitUI.show()  # Show the UI
-	elif DOCK_WITH_NUKE_UI:
-		prefix = ''
-		basename = os.path.basename(__file__)
-		module_name = basename[: basename.rfind('.')]
-		if __name__ == module_name:
-			prefix = module_name + '.'
-		panel = nukescripts.panels.registerWidgetAsPanel(
-			widget=prefix + WINDOW_TITLE,  # module_name.Class_name
-			name=WINDOW_TITLE,
-			id='uk.co.thefoundry.' + WINDOW_TITLE,
-			create=True)
-		pane = nuke.getPaneFor('Properties.1')
-		panel.addToPane(pane)
-		renderSubmitUI = panel.customKnob.getObject().widget
-		_nuke_set_zero_margins(renderSubmitUI)
+	renderSubmitUI = RenderSubmitUI(parent=_nuke_main_window())
+
+	# renderSubmitUI.setWindowFlags(QtCore.Qt.Tool)
+
+	renderSubmitUI.display(**kwargs)  # Show the UI
 
 
 # Detect environment and run application
 if os.environ['IC_ENV'] == 'MAYA':
 	import maya.cmds as mc
-	run_maya()
+	verbose.print_("GPS Render Submit for Maya", 4)
+	# run_maya()
 elif os.environ['IC_ENV'] == 'NUKE':
 	import nuke
 	import nukescripts
-	run_nuke()
+	verbose.print_("GPS Render Submit for Nuke", 4)
+	# run_nuke()
 # elif __name__ == '__main__':
 # 	run_standalone()