+++ rsc/nuke/nukeOps.py
@@ -97,15 +97,14 @@ def submitRender():
 		render license.
 	"""
 	import render_submit
-	# try:
-	# 	renderSubmitDialog.display(flags='-i')
-	# except AttributeError:
-	renderSubmitDialog = render_submit.renderSubmitDialog()
-	renderSubmitDialog.display(flags='-i')
+	render_submit.run_nuke(flags='-i')
+	# renderSubmitDialog = render_submit.renderSubmitDialog()
+	# renderSubmitDialog.display(flags='-i')
 
 
 def submitRenderSelected():
-	""" Launches GPS Render Submitter window, for rendering the currently selected write nodes only.
+	""" Launches GPS Render Submitter window, for rendering the currently
+		selected write nodes only.
 	"""
 	writeNodes = ''
 	selectedNodes = nuke.selectedNodes()
@@ -118,11 +117,9 @@ def submitRenderSelected():
 	# 	writeNodes = '-X %s ' %selectedNode.name()
 
 	import render_submit
-	# try:
-	# 	renderSubmitDialog.display(flags='-i %s' %writeNodes)
-	# except AttributeError:
-	renderSubmitDialog = render_submit.renderSubmitDialog()
-	renderSubmitDialog.display(flags='-i %s' %writeNodes)
+	render_submit.run_nuke(flags='-i %s' %writeNodes)
+	# renderSubmitDialog = render_submit.renderSubmitDialog()
+	# renderSubmitDialog.display(flags='-i %s' %writeNodes)
 
 
 def viewerSnapshot(pblPath):